+++ packages/imperative/src/profiles/index.ts
@@ -22,7 +22,6 @@ export * from "./src/doc/api/IProfileManagerFactory";
 
 export * from "./src/doc/parms/ILoadProfile";
 export * from "./src/doc/parms/IProfileManager";
-export * from "./src/doc/parms/ISetDefaultProfile";
 export * from "./src/doc/parms/IValidateProfile";
 export * from "./src/doc/parms/IValidateProfileForCLI";
 export * from "./src/doc/parms/IValidateProfileWithSchema";