+++ packages/imperative/src/profiles/src/doc/parms/index.ts
@@ -12,7 +12,6 @@
 export * from "./ILoadAllProfiles";
 export * from "./ILoadProfile";
 export * from "./IProfileManager";
-export * from "./ISetDefaultProfile";
 export * from "./IValidateProfile";
 export * from "./IValidateProfileForCLI";
 export * from "./IValidateProfileWithSchema";